+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This section contains changes that have been committed but not yet released.
 
 ### Added
 
+* Added support for provider detection
+* Added an enum for all known providers for native authentication
+
 ### Changed
 
 ### Deprecated